Fix DOIs in publication data file (#15)
In .bib files the underscores in DOIs are usually escaped ('\_').
When converting biblatex to YAML, pandoc escapes both characters
again.  This results in '\\\_'.  When writing the DOI url, this
is repeated to produced '\\\\\\_'.

I couldn't find a way to turn this of.  Instead this commit extends
the generation python script to remove those superfluous escapes.